Move sim_t::main() from sc_sim.cpp to sc_main.cpp. Semantically, it's part of the CLI controller and not part of the engine; there's no reason to link it into the GUI.
git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@15058 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/sc_action.cpp
@@ -1873,10 +1873,10 @@ int action_t::hasted_num_ticks( double haste, timespan_t d )
 {
   if ( ! hasted_ticks )
   {
-    if ( d < timespan_t::zero() ) 
+    if ( d < timespan_t::zero() )
       return num_ticks;
     else
-      return d / base_tick_time;
+      return static_cast<int>( d / base_tick_time );
   }
 
 #ifndef NDEBUG
